+++ src/cp1610.test.ts
@@ -2,35 +2,12 @@ import {describe, expect, test} from "vitest";
 import fs from "fs";
 import glob from "glob-promise";
 
-import {BusDevice, CP1610, forTestSuite} from "./cp1610";
-import { ROM } from "./ROM";
-import { RAM } from "./RAM";
-import { Bus } from "./Bus";
+import {ROM} from "./ROM";
+import {RAM} from "./RAM";
+import {Bus, BusDevice} from "./Bus";
 import {UnreachableCaseError} from "./UnreachableCaseError";
-
-const {decodeOpcode} = forTestSuite;
-
-describe("decodeOpcode", () => {
-  const tests = [
-    {addrs: [0x0000], mnemonic: "HLT"},
-    {addrs: [0x0001], mnemonic: "SDBD"},
-    {addrs: [0x0004], mnemonic: "J"},
-    {
-      addrs: [0x0200, 0x222, 0x23f],
-      mnemonic: "B",
-    },
-  ] as const;
-  for (const {addrs, mnemonic} of tests) {
-    test(mnemonic, () => {
-      for (const addr of addrs) {
-        const instructionConfig = decodeOpcode(addr);
-        expect(instructionConfig).toBeDefined();
-        expect(instructionConfig).toBeDefined();
-        expect(instructionConfig?.mnemonic).toBe(mnemonic);
-      }
-    });
-  }
-});
+import {CP1610} from "./cp1610";
+import {decodeOpcode} from "./decodeOpcode";
 
 const readRomIntoUint16Array = (path: string) => {
   const romBuffer = fs.readFileSync(path);
@@ -41,136 +18,6 @@ const readRomIntoUint16Array = (path: string) => {
   return romData;
 };
 
-describe("bus devices", () => {
-  test("basic instruction fetching and jumping", () => {
-    const bus = new Bus();
-    const cpu = new CP1610(bus);
-    const devices = [
-      cpu,
-      // Rough approximation of various RAM devices:
-      new RAM(bus, 0x0000, 0x1000),
-      // Rough approximation of EXEC ROM
-      new ROM(bus, 0x1000, readRomIntoUint16Array("./roms/exec.bin")),
-    ];
-    const tick = () => devices.forEach((device) => device.clock());
-    const microCycle = () => {
-      tick();
-      tick();
-      tick();
-      tick();
-    };
-
-    const step = () => {
-      while (cpu.state !== "FETCH_OPCODE:BAR") {
-        tick();
-      }
-    };
-
-    // Initialization sequence; CPU should first initialize its PC to $1000:
-    microCycle();
-    expect(cpu.r[7]).toEqual(0x1000);
-
-    // Wait one NACT cycle:
-    expect(bus.flags).toBe(Bus.___);
-    microCycle();
-    // Begin opcode fetch; CPU should assert PC to bus:
-    expect(bus.flags).toBe(Bus.BAR);
-    microCycle();
-    expect(bus.data).toBe(0x1000);
-    expect(cpu.r[7]).toBe(0x1001);
-    // NACT:
-    expect(bus.flags).toBe(Bus.___);
-    microCycle();
-    // Now the ROM should assert data at $1000 to bus:
-    expect(bus.flags).toBe(Bus.DTB);
-    microCycle();
-    expect(bus.data).toBe(0x0004);
-    // ...and the CPU should have read the bus data into its next opcode
-    expect(cpu.opcode).toBe(0x0004);
-
-    // While in NACT, CPU then decodes the opcode into an instruction
-    expect(bus.flags).toBe(Bus.___);
-    microCycle();
-
-    // ...which should be JUMP instruction
-    expect(cpu.instruction?.mnemonic).toBe("J");
-
-    // Now we should read two more decles as arguments to the Jump instruction
-
-    // Read first argument...
-    {
-      expect(bus.flags).toBe(Bus.BAR);
-      microCycle();
-      expect(bus.data).toBe(0x1001);
-
-      expect(bus.flags).toBe(Bus.___);
-      microCycle();
-
-      expect(bus.flags).toBe(Bus.DTB);
-      microCycle();
-      expect(bus.data).toBe(0x0112);
-      // First argument read:
-      expect(cpu.args[0]).toBe(0x0112);
-
-      expect(bus.flags).toBe(Bus.___);
-      microCycle();
-    }
-
-    // Read second argument...
-    {
-      expect(bus.flags).toBe(Bus.BAR);
-      microCycle();
-      expect(bus.data).toBe(0x1002);
-
-      expect(bus.flags).toBe(Bus.___);
-      microCycle();
-
-      expect(bus.flags).toBe(Bus.DTB);
-      microCycle();
-      expect(bus.data).toBe(0x0026);
-      // Second argument read:
-      expect(cpu.args[1]).toBe(0x0026);
-
-      expect(bus.flags).toBe(Bus.___);
-      microCycle();
-    }
-
-    // Skip over the rest of the NACTs:
-    step();
-
-    expect(cpu.r[7]).toBe(0x1026);
-    // JSRD should disable Interrupt Enable flag
-    expect(cpu.i).toBe(false);
-    // Return address should be stored in R5
-    expect(cpu.r[5]).toBe(0x1003);
-
-    expect(bus.flags).toBe(Bus.___);
-    microCycle();
-
-    // Begin opcode fetch; CPU should assert PC to bus:
-    expect(bus.flags).toBe(Bus.BAR);
-    microCycle();
-    expect(bus.data).toBe(0x1026);
-    expect(cpu.r[7]).toBe(0x1027);
-    // NACT:
-    expect(bus.flags).toBe(Bus.___);
-    microCycle();
-    // Now the ROM should assert data at $1026 to bus:
-    expect(bus.flags).toBe(Bus.DTB);
-    microCycle();
-    expect(bus.data).toBe(0x02be);
-    // ...and the CPU should have read the bus data into its next opcode
-    expect(cpu.opcode).toBe(0x02be);
-
-    // While in NACT, CPU then decodes the opcode into an instruction
-    expect(bus.flags).toBe(Bus.___);
-    microCycle();
-
-    // ...which should be MVII instruction
-    expect(cpu.instruction?.mnemonic).toBe("MVII");
-  });
-});
-
 const $word = (n: number) => "$" + word(n);
 const word = (n: number) => n.toString(16).toUpperCase().padStart(4, "0");
 
@@ -180,7 +27,6 @@ describe("jzIntv fixtures", async () => {
   // module.
 
   const fixtures = await glob("./src/fixtures/*.jzintv.txt");
-  console.log("fixtures", fixtures);
   for (const fixturePath of fixtures) {
     test(fixturePath, async () => {
       class BusSniffer implements BusDevice {
@@ -188,7 +34,6 @@ describe("jzIntv fixtures", async () => {
         ticks: number = 0;
         addr: number = 0xffff;
         data: number = 0xffff;
-        log: string[] = [];
         /**
          * HACK
          *
@@ -201,8 +46,11 @@ describe("jzIntv fixtures", async () => {
          * as it seems to lead to valid output.
          */
         seenReads = new Set<number>();
-        constructor(bus: Bus) {
+        onBusActivity: (logLine: string) => void;
+
+        constructor(bus: Bus, onBusActivity: (logLine: string) => void) {
           this.bus = bus;
+          this.onBusActivity = onBusActivity;
         }
 
         clock(): void {
@@ -231,7 +79,18 @@ describe("jzIntv fixtures", async () => {
               // The CPU asserts nothing during this phase -- rather, it expects the
               // currently addressed device to inform the rest of the machine of the
               // address for the next access.
-              if (this.ticks === 3) this.addr = this.bus.data;
+              if (this.ticks === 3) {
+                if (!this.seenReads.has(this.addr)) {
+                  this.seenReads.add(this.addr);
+                  this.onBusActivity(
+                    `RD a=${$word(this.addr)} d=${word(
+                      this.bus.data,
+                    )} CP-1610 (PC = ${$word(cpu.r[7])}) t=${stepCycle}`,
+                  );
+                  // console.error(log.at(-1), cpu.busSequence, cpu.busSequenceIndex, this.bus.toString())
+                }
+                this.addr = this.bus.data;
+              }
               return;
             }
             case Bus.DTB: {
@@ -242,11 +101,17 @@ describe("jzIntv fixtures", async () => {
                 this.data = this.bus.data;
                 if (!this.seenReads.has(this.addr)) {
                   this.seenReads.add(this.addr);
-                  this.log.push(
-                    `RD a=${$word(this.addr)} d=${word(this.data)} ${
-                      cpu.state
-                    }`,
+                  this.onBusActivity(
+                    `RD a=${$word(this.addr)} d=${word(
+                      this.data,
+                    )} CP-1610 (PC = ${$word(cpu.r[7])}) t=${stepCycle}`,
                   );
+                  // console.error(
+                  //   log.at(-1),
+                  //   cpu.busSequence,
+                  //   cpu.busSequenceIndex,
+                  //   this.bus.toString(),
+                  // );
                 }
               }
               return;
@@ -263,8 +128,10 @@ describe("jzIntv fixtures", async () => {
               // data.
               if (this.ticks === 3) {
                 this.data = this.bus.data;
-                this.log.push(
-                  `WR a=${$word(this.addr)} d=${word(this.data)} ${cpu.state}`,
+                this.onBusActivity(
+                  `WR a=${$word(this.addr)} d=${word(
+                    this.data,
+                  )} CP-1610 (PC = ${$word(cpu.r[7])}) t=${stepCycle}`,
                 );
               }
               return;
@@ -304,12 +171,12 @@ describe("jzIntv fixtures", async () => {
           }
         }
 
-        debug_read(addr: number): number | null {
-          throw new Error("Method not implemented.");
+        debug_read(_addr: number): number | null {
+          return null;
         }
       }
 
-      const lines = (
+      const expectedLog = (
         await fs.promises.readFile(fixturePath, {encoding: "utf-8"})
       ).split("\n");
       const romPath =
@@ -317,8 +184,8 @@ describe("jzIntv fixtures", async () => {
 
       const bus = new Bus();
       const cpu = new CP1610(bus);
-      const busSniffer = new BusSniffer(bus);
-
+      const log: string[] = [];
+      const busSniffer = new BusSniffer(bus, (busLog) => log.push(busLog));
       const devices = [
         cpu,
         // Rough approximation of various RAM devices:
@@ -331,15 +198,20 @@ describe("jzIntv fixtures", async () => {
       let cycles = 0;
       const tick = () => {
         cycles += 1;
+        bus.clock();
         devices.forEach((device) => device.clock());
       };
 
+      let stepCycle = 0;
       const step = () => {
-        while (cpu.state === "FETCH_OPCODE:BAR") {
+        stepCycle = Math.max(0, cycles) / 4;
+        log.push(cpuStatus());
+
+        while (cpu.busSequence === "INSTRUCTION_FETCH") {
           tick();
         }
         // @ts-ignore
-        while (cpu.state !== "FETCH_OPCODE:BAR") {
+        while (cpu.busSequence !== "INSTRUCTION_FETCH") {
           tick();
         }
       };
@@ -351,18 +223,24 @@ describe("jzIntv fixtures", async () => {
         }
         return 0xffff;
       };
+      const peekBusSDBD = (addr: number) => {
+        return ((peekBus(addr + 1) & 0x00ff) << 8) | (peekBus(addr) & 0x00ff);
+      };
 
-      const cpuFlags = () =>
-        [
+      const cpuFlags = () => {
+        let prevInstruction = cycles / 4 > 13 ? decodeOpcode(cpu.opcode) : null;
+
+        return [
           cpu.s ? "S" : "-",
           cpu.z ? "Z" : "-",
           cpu.o ? "O" : "-",
           cpu.c ? "C" : "-",
           cpu.i ? "I" : "-",
           cpu.d ? "D" : "-",
-          cpu.prevInstruction?.interruptable ? "i" : "-",
+          prevInstruction?.interruptable ? "i" : "-",
           "-", // TODO: interrupt state info
         ].join("");
+      };
 
       const cpuCurrentInstruction = () => {
         const pc = cpu.r[7];
@@ -427,9 +305,11 @@ describe("jzIntv fixtures", async () => {
           case "MVII": {
             if (reg0Index === 6) {
               return `PULR R${reg1Index}`;
-            } else {
-              const data = peekBus(pc + 1);
+            } else if (reg0Index === 7) {
+              const data = cpu.d ? peekBusSDBD(pc + 1) : peekBus(pc + 1);
               return `MVII #${$word(data)},R${reg1Index}`;
+            } else {
+              return `MVI@ R${reg0Index},R${reg1Index}`;
             }
           }
           case "MVO":
@@ -442,6 +322,14 @@ describe("jzIntv fixtures", async () => {
               return `PSHR R${reg1Index}`;
             } else {
               return `${instruction.mnemonic} R${reg1Index},R${reg0Index}`;
+            }
+          }
+
+          case "MOVR": {
+            if (reg1Index === 7) {
+              return `JR,R${reg0Index}`;
+            } else {
+              return `${instruction.mnemonic} R${reg0Index},R${reg1Index}`;
             }
           }
 
@@ -485,6 +373,13 @@ describe("jzIntv fixtures", async () => {
             })();
             return `${mnemonic} ${$word(addr)}`;
           }
+          case "SWAP": {
+            return `SWAP R${reg1Index}`;
+          }
+          case "ANDI": {
+            const data = cpu.d ? peekBusSDBD(pc + 1) : peekBus(pc + 1);
+            return `${instruction.mnemonic} #${$word(data)},R${reg1Index}`;
+          }
         }
 
         return instruction.mnemonic;
@@ -494,25 +389,30 @@ describe("jzIntv fixtures", async () => {
         const reg = (n: number) => word(n);
         return `${[...cpu.r]
           .map(reg)
-          .join(" ")} ${cpuFlags()}  ${cpuCurrentInstruction()}|${cycles / 4}`;
-      };
-
-      const busStatus = (): string => {
-        const [lastLog] = busSniffer.log.splice(0, 1);
-        if (lastLog == null) return "BUS INACTIVE";
-        return lastLog;
+          .join(" ")} ${cpuFlags()}  ${cpuCurrentInstruction()}|${
+          // HACKish cycle count offset
+          stepCycle
+        }`;
       };
 
       const normalizeCpuStatus = (line: string) =>
         line
           .replace(/  +/g, "|")
-          // TODO: Do we want to try and match cycle count of jzIntv?
           .split("|")
+          // TODO: It's unclear "when" CPU status is logged in jzIntv so I'm
+          // disabling timing comparisons, for now.
           .slice(0, -1)
           .join(",");
 
-      const normalizeBusStatus = (line: string) =>
-        line.split(" ").slice(0, 3).join(" ");
+      const normalizeBusStatus = (line: string) => {
+        const s = line.split(/ +/g);
+
+        // TODO: It's unclear "when" CPU status is logged in jzIntv so I'm
+        // disabling timing comparisons, for now.
+        // return `${s[0]} ${s[1]} ${s[2]} ${s[7]}`;
+        return `${s[0]} ${s[1]} ${s[2]}`;
+      };
+
       const normalizeLine = (line: string) => {
         if (line.startsWith("RD") || line.startsWith("WR")) {
           return normalizeBusStatus(line);
@@ -520,40 +420,29 @@ describe("jzIntv fixtures", async () => {
           return normalizeCpuStatus(line);
         }
       };
+
       // Hackish: Allow reset sequence to do its thing
       step();
-      expect(cpu.state).toMatchInlineSnapshot(`"FETCH_OPCODE:BAR"`);
+
       // Reset cycle count after reset:
       cycles = 0;
+      log.length = 0;
 
-      expect(lines.length).toBeGreaterThan(0);
-      let prevLines: string[] = [];
-      for (let lineNumber = 0; lineNumber < lines.length; ++lineNumber) {
-        // First account for any bus activity:
-        for (const busLogLine of busSniffer.log) {
-          console.log(busLogLine);
-          const line_ = lines[lineNumber]!;
-          const prefix = `${(lineNumber + 1).toString(10).padStart(4, " ")}: `;
-          const line = prefix + normalizeLine(line_);
-          const prev = prevLines.slice(-20).join("\n");
-          expect(
-            (prev ? prev + "\n" : "") + prefix + normalizeBusStatus(busLogLine),
-          ).toEqual((prev ? prev + "\n" : "") + line);
-          prevLines.push(line);
-          lineNumber += 1;
-        }
-        // Reset bus activity log
-        busSniffer.log.length = 0;
+      expect(expectedLog.length).toBeGreaterThan(0);
 
-        const line_ = lines[lineNumber]!;
-        const prefix = `${(lineNumber + 1).toString(10).padStart(4, " ")}: `;
-        const line = prefix + normalizeCpuStatus(line_);
-        const prev = prevLines.slice(-20).join("\n");
-        expect(
-          (prev ? prev + "\n" : "") + prefix + normalizeCpuStatus(cpuStatus()),
-        ).toEqual((prev ? prev + "\n" : "") + line);
+      while (log.length < expectedLog.length) {
         step();
-        prevLines.push(line);
+      }
+
+      const normalizedLog = log.map(normalizeLine);
+      const normalizedExpectedLog = expectedLog.map(normalizeLine);
+      for (let i = 0; i < normalizedLog.length; ++i) {
+        const prefixLine = (line: string, j: number) =>
+          `${(j + 1).toString(10).padStart(4, " ")}: ` + line;
+
+        expect(normalizedLog.slice(0, i).map(prefixLine).join("\n")).toEqual(
+          normalizedExpectedLog.slice(0, i).map(prefixLine).join("\n"),
+        );
       }
     });
   }
